% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_distribution_age.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Calcul distribution âge}
\usage{
calcul_distribution_age(df)
}
\arguments{
\item{df}{Un dataframe conforme au schéma.}
}
\value{
Un vecteur numérique contenant les quantiles (0%, 25%, 50%, 75%, 100%) de la distribution des âges.
}
\description{
Calcule la distribution des âges dans un df
}
\details{
La fonction valide le schéma du dataframe, convertit les dates de naissance en format date,
calcule l'âge des individus et retourne les quantiles de la distribution des âges.
Avertissement si des dates de naissance invalides sont détectées.

Cette fonction utilise le package `stats` pour calculer les quantiles
}
\examples{
df_exemple <- data.frame(
  Code.du.département = rep("", 4),
  Libellé.du.département = rep("", 4),
  Code.de.la.collectivité.à.statut.particulier = rep("", 4),
  Libellé.de.la.collectivité.à.statut.particulier = rep("", 4),
  Code.de.la.commune = rep("", 4),
  Libellé.de.la.commune = rep("", 4),
  Nom.de.l.élu = rep("", 4),
  Prénom.de.l.élu = rep("", 4),
  Code.sexe = rep("", 4),
  Date.de.naissance = c("01-01-1950", "15-06-1970", "10-09-1985", "pas une date"),
  Code.de.la.catégorie.socio.professionnelle = rep("", 4),
  Libellé.de.la.catégorie.socio.professionnelle = rep("", 4),
  Date.de.début.du.mandat = rep("", 4),
  Libellé.de.la.fonction = rep("", 4),
  Date.de.début.de.la.fonction = rep("", 4),
  Code.nationalité = rep("", 4)
)
calcul_distribution_age(df_exemple)

}
