% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Générer rapport}
\usage{
generer_rapport(code_commune, code_departement, output)
}
\arguments{
\item{output}{Le nom de sortie du fichier pdf généré, sous forme de chaîne de caractères.}

\item{commune}{Un code de commune sous forme de chaîne de caractères (ex : "44109").}

\item{departement}{Un code de département sous forme de chaîne de caractères (ex : "44").}
}
\value{
Aucun retour explicite. Un fichier PDF
}
\description{
Cette fonction génère un rapport PDF se basant sur un modèle quarto stocké dans le dossier 'inst/' du package.
Elle utilise les codes de la commune et du département donnée, et sauvegarde le rapport en .pdf à la racine.
}
\examples{
\dontrun{
generer_rapport(
  commune = "44109",
  departement = "44",
  output = "rapport_nantes.pdf"
)
}

}
