% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code_professions.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Plot code professions}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{Un dataframe conforme au schéma.}
}
\value{
Un barplot du nombre d'élus par CSP.
}
\description{
Visualise le nombre d'élus par catégorie socio-professionnelle (CSP)
}
\details{
Cette fonction prend un dataframe contenant des informations sur les élus et
génère un graphique à barres représentant le nombre d'élus pour chaque catégorie
socio-professionnelle, trié par nombre décroissant.

Cette fonction utilise le package `dplyr` pour manipuler le dataframe et le package `ggplot2` pour la visualisation des données.
}
\examples{
df_exemple <- data.frame(
  Code.du.département = rep("", 4),
  Libellé.du.département = rep("", 4),
  Code.de.la.collectivité.à.statut.particulier = rep("", 4),
  Libellé.de.la.collectivité.à.statut.particulier = rep("", 4),
  Code.de.la.commune = rep("", 4),
  Libellé.de.la.commune = rep("", 4),
  Nom.de.l.élu = rep("", 4),
  Prénom.de.l.élu = rep("", 4),
  Code.sexe = rep("", 4),
  Date.de.naissance = c("", 4),
  Code.de.la.catégorie.socio.professionnelle = c("A", "B", "C", "A", "B", "A"),
  Libellé.de.la.catégorie.socio.professionnelle = rep("", 4),
  Date.de.début.du.mandat = rep("", 4),
  Libellé.de.la.fonction = rep("", 4),
  Date.de.début.de.la.fonction = rep("", 4),
  Code.nationalité = rep("", 4)
)
plot_code_professions(df_exemple)

}
