% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_departement.R
\name{plot_departement}
\alias{plot_departement}
\title{Plot departement}
\usage{
plot_departement(df)
}
\arguments{
\item{df}{Un dataframe d'un unique département et conforme au schéma validé
par`validate_schema`}
}
\value{
Un diagramme en barres des catégories socio-professionnelles pour la commune donnée.
}
\description{
Cette fonction génère un graphique (barplot) pour chaque code professionnel et le nombre
d’élus ayant ce code.
}
\details{
Cette fonction utilise le package `dplyr` pour manipuler le dataframe et le package `ggplot2` pour la visualisation des données.
}
\examples{
df_exemple <- data.frame(
  Code.du.département = rep("", 4),
  Libellé.du.département = rep("", 4),
  Code.de.la.collectivité.à.statut.particulier = rep("", 4),
  Libellé.de.la.collectivité.à.statut.particulier = rep("", 4),
  Code.de.la.commune = rep("", 4),
  Libellé.de.la.commune = rep("", 4),
  Nom.de.l.élu = rep("", 4),
  Prénom.de.l.élu = rep("", 4),
  Code.sexe = rep("", 4),
  Date.de.naissance = c("", 4),
  Code.de.la.catégorie.socio.professionnelle = rep("", 4),
  Libellé.de.la.catégorie.socio.professionnelle = rep("", 4),
  Date.de.début.du.mandat = rep("", 4),
  Libellé.de.la.fonction = rep("", 4),
  Date.de.début.de.la.fonction = rep("", 4),
  Code.nationalité = rep("", 4)
)
plot_departement(df_exemple)

}
