% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_commune.R
\name{summary_commune}
\alias{summary_commune}
\title{summary commune}
\usage{
summary_commune(x)
}
\arguments{
\item{x}{Un dataframe conforme au schéma.
La classe de l'objet df sera modifié en "commune" si nécessaire.}
}
\value{
La fonction affiche directement un résumé dans la console (utilisation de `cat`)
}
\description{
Cette fonction génère un résumé des informations relatives à une commune à partir d'un dataframe
contenant des informations sur les élus. Elle affiche le nom de la commune, le nombre d'élus,
la distribution des âges des élus et les détails de l'élu le plus âgé.
}
\examples{
df_exemple <- data.frame(
  Code.du.département = rep("", 4),
  Libellé.du.département = rep("", 4),
  Code.de.la.collectivité.à.statut.particulier = rep("", 4),
  Libellé.de.la.collectivité.à.statut.particulier = rep("", 4),
  Code.de.la.commune = c("1","1","1","1"),
  Libellé.de.la.commune = c("A","A","A","A"),
  Nom.de.l.élu = rep("", 4),
  Prénom.de.l.élu = rep("", 4),
  Code.sexe = rep("", 4),
  Date.de.naissance = c("", 4),
  Code.de.la.catégorie.socio.professionnelle = c("A", "B", "C", "A", "B", "A"),
  Libellé.de.la.catégorie.socio.professionnelle = rep("", 4),
  Date.de.début.du.mandat = rep("", 4),
  Libellé.de.la.fonction = rep("", 4),
  Date.de.début.de.la.fonction = rep("", 4),
  Code.nationalité = rep("", 4)
)
summary_commune(df_exemple)

}
