% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_departement.R
\name{summary_departement}
\alias{summary_departement}
\title{summary departement}
\usage{
summary_departement(x)
}
\arguments{
\item{x}{Un dataframe conforme au schéma.
La classe de l'objet df sera modifié en "departement" si nécessaire.}
}
\value{
La fonction affiche directement un résumé dans la console (utilisation de `cat`)
}
\description{
Cette fonction génère un résumé des informations relatives à un département à partir d'un dataframe contenant des informations sur les élus.
Elle affiche le nom du département, le nombre de communes, d'élus et la distribution d'âge des élus dans celui ci. La fonction
affiche aussi les informations de l'élu le plus âgée et le plus jeune du département ainsi que les statistiques de la commune
avec la moyenne d'âge la plus faible et la plus élevée.
}
\details{
Cette fonction utilise le package `dplyr` et `lubridate` pour manipuler le dataframe.
}
\examples{
df_exemple <- data.frame(
  Code.du.département = c("1","1","1","1"),
  Libellé.du.département = c("A","A","A","A"),
  Code.de.la.collectivité.à.statut.particulier = rep("", 4),
  Libellé.de.la.collectivité.à.statut.particulier = rep("", 4),
  Code.de.la.commune = rep("", 4),
  Libellé.de.la.commune = rep("", 4),
  Nom.de.l.élu = rep("", 4),
  Prénom.de.l.élu = rep("", 4),
  Code.sexe = rep("", 4),
  Date.de.naissance = c("", 4),
  Code.de.la.catégorie.socio.professionnelle = rep("", 4),
  Libellé.de.la.catégorie.socio.professionnelle = rep("", 4),
  Date.de.début.du.mandat = rep("", 4),
  Libellé.de.la.fonction = rep("", 4),
  Date.de.début.de.la.fonction = rep("", 4),
  Code.nationalité = rep("", 4)
)
summary.departement(df_exemple)

}
