% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_schema.R
\name{validate_schema}
\alias{validate_schema}
\title{Validate schema}
\usage{
validate_schema(df)
}
\arguments{
\item{df}{Un dataframe à valider}
}
\value{
Aucun retour explicite. Erreur si le schéma ne correpond pas.
}
\description{
Vérifie la conformité du schéma d'un df
}
\details{
La fonction valide si un df possède exactement les colonnes attendues.
Si le schéma ne correspond pas, une erreur est levée.
}
\examples{
df_bon <- data.frame(
  Code.du.département = double(),
  Libellé.du.département = character(),
  Code.de.la.collectivité.à.statut.particulier = double(),
  Libellé.de.la.collectivité.à.statut.particulier = character(),
  Code.de.la.commune = double(),
  Libellé.de.la.commune = character(),
  Nom.de.l.élu = character(),
  Prénom.de.l.élu = character(),
  Code.sexe = character(),
  Date.de.naissance = character(),
  Code.de.la.catégorie.socio.professionnelle = double(),
  Libellé.de.la.catégorie.socio.professionnelle = character(),
  Date.de.début.du.mandat = character(),
  Libellé.de.la.fonction = character(),
  Date.de.début.de.la.fonction = character(),
  Code.nationalité = character()
)

validate_schema(df_bon) #Pas d'erreur

df_pas_bon <- data.frame(
département = double(),
csp = character(),
A = 1:5,
B = 10:12
)

valide_schema(df_pas_bon) #Erreur : Le schéma des colonnes du dataframe ne correspond pas au format attendu.
}
